[FEAT] 콘텐츠 생성, 콘텐츠 카테고리 수정 API 유저 검증 로직 추가, 생성 API statusCode 201로 수정

+++ functions/api/routes/content/contentCategoryPATCH.js
@@ -4,7 +4,8 @@ const util = require('../../../lib/util');
 const statusCode = require('../../../constants/statusCode');
 const responseMessage = require('../../../constants/responseMessage');
 const db = require('../../../db/db');
-const { categoryContentDB } = require('../../../db');
+const { categoryDB, categoryContentDB } = require('../../../db');
+const e = require('cors');
 
 /**
  *  @route PATCH /content/category
@@ -26,14 +27,28 @@ module.exports = async (req, res) => {
   try {
     client = await db.connect(req);
 
-    const deletedCategoryIds = await categoryContentDB.deleteCategoryContentByContentId(client, contentId); // 기존 카테고리-콘텐츠 관계 모두 삭제
-
+    let flag = true; // flag 변수 결과에 따라 categoryContent를 추가할 지, 에러를 보낼 지 결정
     for (const newCategoryId of newCategoryIds) {
+      // 카테고리 배열의 id 중 하나라도 유저의 카테고리가 아닐 경우, 에러 전송
+      const newCategory = await categoryDB.getCategory(client, newCategoryId);
+      if (!newCategory || newCategory.userId !== userId) {
+        // 카테고리가 아예 존재하지 않거나, 해당 유저의 카테고리가 아닌 경우
+        flag = false;
+      }
+    }
+
+    if (flag) {
+      // 유저가 해당 카테고리를 가지고 있을 때
+      const deletedCategoryIds = await categoryContentDB.deleteCategoryContentByContentId(client, contentId); // 기존 카테고리-콘텐츠 관계 모두 삭제
+      for (const newCategoryId of newCategoryIds) {
         // 새로운 카테고리마다 새로운 카테고리-콘텐츠 관계 생성
         const newCategoryContent = await categoryContentDB.addCategoryContent(client, newCategoryId, contentId);
-    }
-    
-    res.status(statusCode.OK).send(util.success(statusCode.OK, responseMessage.CHANGE_CONTENT_CATEGORY_SUCCESS));
+      }
+      res.status(statusCode.OK).send(util.success(statusCode.OK, responseMessage.CHANGE_CONTENT_CATEGORY_SUCCESS));    
+    } else {
+      // 카테고리를 변경할 수 없는 경우, 에러 전송
+      res.status(statusCode.NOT_FOUND).send(util.fail(statusCode.NOT_FOUND, responseMessage.NO_CATEGORY));
+    }   
     
   } catch (error) {
     functions.logger.error(`[ERROR] [${req.method.toUpperCase()}] ${req.originalUrl}`, `[CONTENT] ${error}`);
